Modules.tcl [BUGFIX]: add APPLICATION_CORE_ENTITY_ONLY as ARCHGRP array field
Previous implementation was focused on simple list but this was
conflicting with the entry conversion from list to array which requires
even number of elements (odd is key, even is value). So
APPLICATION_CORE_ENTITY_ONLY was converted from simple string to boolean
variable which is handed through ARCHGRP. Because this file is used in
the case when only APPLICATION_CORE entity is present in the design, the
initialization of DMA_BUS_ARCH was moved to the else branch of the
conditional statement.